feat: 클라이언트타입에 Kakao 추가 - #64
